`tests/unit/config/__init__.py` -> `tests/unit/config/test_config.py`
No top level `__init__.py` should contain tests
